third_party/blink/web_tests/fast/js/Promise-native-reject-expected.txt

Test Promise reject inside Blink.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS result is 42
PASS successfullyParsed is true

TEST COMPLETE
